Bigger update this time. Following features were implemented -categories -favourites -userlistings
implemented error messages to login and newpost, but must remake them
to be automatic on the form so the user doesnt have to type the texts allover
again. Made the navbar move to the pages it says and made the backend search
for the favourites and posts the users has that is logged in. Must still clear
and refractor all of the code and the main thing that i want working is the
automatic error messages to the forms working like they should be and
also the search.